makeRequest is now context aware
makeRequest receives an aditional parameter
called 'type' in which indicates the current
action being performed.

Since makeRequest is context aware, if the user
selects 'Next' and it fails on the first try,
it will continue to request three more times
but instead of going back to the day before,
it will continue on forward to the next day
and so forth.