Add new ActionTypes to prevent repeated requests
Add two new ActionTypes:

- NEWEST: Used only once, the first time the app requests.
- LATEST: Used in case the newest fails the request.

In the first time that the App executes it will use the action type
NEWEST, it make a request using the current date and set the
action type to LATEST in case occurs a failure in the request.
Subsequently, LATEST sets the current date to a day before and
updates the LATEST_DAY constant to this new date, it also uses
this date to make the new request.

Also includes:

- random date renamed to randate
- minor change in the css
- a new constant to choose to wait image to download